feat: add ordering to published posts retrieval and improve avatar deletion logic

+++ src/api/users/users.service.ts
@@ -136,9 +136,11 @@ export class UsersService {
       avatar.buffer,
     );
 
-    const oldAvatarKey = user.avatar.split('/').pop();
-    if (oldAvatarKey) {
-      await this.s3Service.deleteFile(oldAvatarKey);
+    if (user.avatar) {
+      const oldAvatarKey = user.avatar.split('/').pop();
+      if (oldAvatarKey) {
+        await this.s3Service.deleteFile(oldAvatarKey);
+      }
     }
 
     const url = `https://${process.env.AWS_S3_BUCKET_NAME}.s3.${process.env.AWS_REGION}.amazonaws.com/${key}`;
